+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 2.0.0
+- Refactor to make it simpler
+- Remove all packages dependencies
+
 ## 1.3.0
 - Refactor the logic related the deleteFirst() in MapStorage
   - as the original method is not necessarily removing the first expired item